debt - remove quickFix/browser
