+++ src/vs/editor/contrib/quickFix/browser/quickFixModel.ts
@@ -13,7 +13,7 @@ import { IMarker, IMarkerService } from 'vs/platform/markers/common/markers';
 import { Range } from 'vs/editor/common/core/range';
 import { ICommonCodeEditor, IPosition, IRange } from 'vs/editor/common/editorCommon';
 import { CodeActionProviderRegistry, CodeAction } from 'vs/editor/common/modes';
-import { getCodeActions } from '../common/quickFix';
+import { getCodeActions } from './quickFix';
 
 
 export class QuickFixOracle {
