+++ src/vs/workbench/test/electron-browser/api/extHostLanguageFeatures.test.ts
@@ -32,7 +32,7 @@ import { getDefinitionsAtPosition, getImplementationsAtPosition, getTypeDefiniti
 import { getHover } from 'vs/editor/contrib/hover/common/hover';
 import { getOccurrencesAtPosition } from 'vs/editor/contrib/wordHighlighter/common/wordHighlighter';
 import { provideReferences } from 'vs/editor/contrib/referenceSearch/browser/referenceSearch';
-import { getCodeActions } from 'vs/editor/contrib/quickFix/common/quickFix';
+import { getCodeActions } from 'vs/editor/contrib/quickFix/browser/quickFix';
 import { getWorkspaceSymbols } from 'vs/workbench/parts/search/common/search';
 import { rename } from 'vs/editor/contrib/rename/browser/rename';
 import { provideSignatureHelp } from 'vs/editor/contrib/parameterHints/common/parameterHints';
